Improve support for work without USB camera HW
In case of absence of USB camera HW we see significant timeout due to
inability of camera frontend to connect to camera_be.
This patch changes behavior of camera_be, so frontend will be able to
connect to camera_be always.
This removes few minutes timeout during start of DomA.

Signed-off-by: Ruslan Shymkevych <ruslan_shymkevych@epam.com>
Reviewed-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>