CameraHandler: Improve "run without hardware" mode
First and foremost:
We shouldn't zero "the whole" response being in that mode,
doing that we are zeroing common for all response types fields:
"id" and "operation" as well which should be a copy of corresponding
fields from the request and seem to be checked back by frontend.
This all leads to the fact that frontend treats these responses as invalid.

What is more that response is already initialized by 0 at the declaration
time in processRequest().

Another reason is the fact that frontend (and/or framework) when requesting
parameters checks them, so we cannot predict what exact values should be
in order to satisfy otherend.

What we can do here is to emulate the absence of controls in a response
for CTRL_ENUM command as we certainly know this is the first command
the frontend usually sends.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>
Reviewed-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>