CameraHandler: Do not call streamStop() with lock held
The problem is that streamStop() needs to wait for eventThread to
finish which in turn might hold the same lock if there is a work
to do, see CommandHandler::onFrameDoneCallback() for details.

If the streamStop() *must* be protected by the lock, we need either
consider dropping the lock from onFrameDoneCallback() or terminate
eventThread forcely.

Signed-off-by: Oleksandr Tyshchenko <oleksandr_tyshchenko@epam.com>