Add support for partial zero-copy
This feature is enabled by default and can be disabled by passing -z
flag to camera_be.

This commit adds support for zero-copy operation for 1 frontend usecase.
This is done by passing the frontend buffers as USERPTR buffers to the
v4l2.

In the case of multiple frontends, some frontend buffers are used as
USERPTR and memcpy is done for the rest of the frontends.

To prevent leaking of one frontend's buffer modifications to another
ones, a strict order of operations is ensured. We first perform all of
the memcopy operations and only release the original buffer after.

In the case of one of the frontends freezing or stopping the stream,
it's buffers are swapped on the fly with the ones from the different
frontend.

Because the Kernel pins the pages used as userptr buffers we need to
wait before releasing them. The waiting is done until one of the
conditions is met:
1. HW stream stopped
2. Same index buffer is overwritten with a different userptr.

This approach ensures that we can always free the buffers in a safe way.

Signed-off-by: Mykyta Poturai <mykyta_poturai@epam.com>